requirements.txt
sentence-transformers
faiss-cpu
pytest
flask
